src/heatmaptile/requirements.txt
Flask
flask
Pillow
psycopg2
python-dotenv